requirements.txt
Flask==2.1.3
numpy==1.23.1
Pillow==8.3.1
requests==2.28.1
torch==1.10.2
torchvision==0.2.2
Werkzeug==2.0.3
validators==0.18.2